[VIEW] Modification du type id de l'objet Event (Long en Integer). Initialisation de la structure pour l'écran des détails d'une session.

+++ jcertif-mobile/project-jcertifmobile-app/src/main/java/com/jcertif/android/JSONHelper.java
@@ -84,7 +84,7 @@ public class JSONHelper {
 
                 Event event = new Event();
 
-                event.id = jsonData.getLong(ID);
+                event.id = jsonData.getInt(ID);
                 event.name = jsonData.getString(NOM);
                 event.description = jsonData.getString(DESCRIPTION);
                 event.keyWord = jsonData.getString(KEY_WORD);
